Updates per Lisa's code review.
