+++ src/js/routes/Guide/UnknownTwitterAccount.jsx
@@ -1,15 +1,11 @@
+import { Button } from "react-bootstrap";
 import React, { Component, PropTypes } from "react";
-import Image from "../../components/Image";
-import { numberWithCommas, removeTwitterNameFromDescription } from "../../utils/textFormat";
+import { Link } from "react-router";
+import TwitterAccountCard from "../../components/Twitter/TwitterAccountCard";
 
 export default class UnknownTwitterAccount extends Component {
   static propTypes = {
-    twitter_handle: PropTypes.string,
-    twitter_description: PropTypes.string,
-    twitter_followers_count: PropTypes.number,
-    twitter_photo_url: PropTypes.string,
-    twitter_user_website: PropTypes.string,
-    twitter_name: PropTypes.string
+    twitter_handle: PropTypes.string
   };
 
   constructor (props) {
@@ -17,54 +13,10 @@ export default class UnknownTwitterAccount extends Component {
   }
 
   render () {
-    let {twitter_handle, twitter_description, twitter_followers_count,
-      twitter_photo_url, twitter_user_website,
-      twitter_name} = this.props;
+    let {twitter_handle} = this.props;
 
-    // If the displayName is in the twitterDescription, remove it from twitterDescription
-    let displayName = twitter_name ? twitter_name : "";
-    let twitterDescription = twitter_description ? twitter_description : "";
-    let twitterDescriptionMinusName = removeTwitterNameFromDescription(displayName, twitterDescription);
-
-    return <div className="card__container">
-      <div className="card__main">
-        <div className="card__media-object">
-          { twitter_photo_url ?
-            <Image imageUrl={twitter_photo_url} class="card__media-object-anchor" /> :
-            <i className="icon-org-lg icon-icon-org-placeholder-6-2 icon-org-resting-color card__media-object-anchor"></i>
-          }
-          <div className="card__media-object-content">
-            <div className="card__display-name">{displayName}</div>
-            { twitterDescriptionMinusName ?
-              <p className="card__description">{twitterDescriptionMinusName}</p> :
-              null
-            }
-            {/* TODO icon-org-placeholder */}
-            { twitter_handle ?
-              <span>@{twitter_handle}&nbsp;&nbsp;</span> :
-              <span></span>
-            }
-            {twitter_followers_count ?
-              <span className="twitter-followers__badge">
-                <span className="fa fa-twitter twitter-followers__icon"></span>
-                {numberWithCommas(twitter_followers_count)}
-              </span> :
-              null
-            }
-            &nbsp;&nbsp;
-            <a href={twitter_user_website} target="_blank">Website</a><br />
-            {/*5 of your friends follow Organization Name<br />*/}
-            {/*
-            <strong>2016 General Election, November 2nd</strong>
-            <br />
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur posuere vulputate massa ut efficitur.
-            Phasellus rhoncus hendrerit ultricies. Fusce hendrerit vel elit et euismod. Etiam bibendum ultricies
-            viverra. Integer ut bibendum ex. Suspendisse eleifend mi accumsan, euismod enim at, malesuada nibh.
-            Duis a eros fringilla, dictum leo vitae, vulputate mi. Nunc vitae neque nec erat fermentum... (more)<br />
-              <AskOrShareAction link_text={"Share Organization"} />*/}
-          </div>
-        </div>
-      </div>
+    return <div>
+      <TwitterAccountCard  {...this.props}/>
     </div>;
   }
 }
